+++ src/calculator/nodes/MultiplyOperator.java
@@ -10,8 +10,8 @@ public class MultiplyOperator extends NodeOperator {
 
     @Override
     public Node interpret() {
-        int leftResult = ((NumberValue)(getLeftParameter().interpret())).getValue();
-        int rightResult = ((NumberValue)getRightParameter().interpret()).getValue();
+        double leftResult = ((NumberValue)(getLeftParameter().interpret())).getValue();
+        double rightResult = ((NumberValue)getRightParameter().interpret()).getValue();
         return new NumberValue(leftResult * rightResult);
     }
 }
